Week7/v1/backend/requirements.txt
flask
flask_sqlalchemy
flasgger
flask_cors
PyJWT>=2.0.0
werkzeug
